app/requirements.txt
exchangelib==4.7.4
pandas==1.4.3
psutil==5.9.1
pyarrow==8.0.0
psycopg2==2.9.3
pyperclip==1.8.2
pywin32==304
PyYAML==6.0
requests==2.28.0
SQLAlchemy==1.4.39
XlsxWriter==3.0.3